mptt-example/requirements-humans.txt
django==1.*
django-extensions
django-mptt
pyreadline # Console tab completion
